Improve positioning elements
