+++ src/index.js
@@ -1,7 +1,7 @@
 import './style.css';
 import renderHeaderFooter from './modules/render_header_footer.js';
+import onLoad from './modules/home_page.js';
 import commentsButtonListener from './modules/popup.js';
-import onLoad from './modules/home.js';
 
 renderHeaderFooter();
 onLoad();
