+++ src/style.css
@@ -41,7 +41,7 @@ footer {
   display: flex;
   flex-direction: column;
   justify-content: center;
-  position: fixed;
+  position: absolute;
   width: 100%;
   height: 10%;
   bottom: 0;
@@ -51,4 +51,23 @@ footer {
 
 #header-logo {
   height: 3rem;
+}
+
+.food-list {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-template-rows: repeat(2);
+  gap: 2vw;
+}
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  height: 80vh;
+  align-items: center;
+  align-self: center;
+}
+
+.card-element {
+  /* height: 15vh; */
 }
