+++ webpack.config.js
@@ -8,7 +8,7 @@ module.exports = {
     index: './src/index.js',
     renderHeaderFooter: './src/modules/render_header_footer.js',
     popup: './src/modules/popup.js',
-    home: './src/modules/home.js',
+    home: './src/modules/home_page.js',
   },
   devtool: 'inline-source-map',
   devServer: {
